Don't require schedule to create event

+++ cgi/admin/add-event.cgi
@@ -17,7 +17,7 @@ my $cgi = CGI->new;
 my $db = db->new();
 
 my $today = strftime("%Y-%m-%d", localtime);
-my $schedule = $cgi->param('csv');
+my $schedule = $cgi->param('csv') or "";
 my $event = $cgi->param('event');
 my $name = $event;
 $name =~ s/^20[0-9]{2}//g;
@@ -36,31 +36,33 @@ sub escapeCsv(){
 
 $webutil->error("Missing event ID") if (!$event);
 $webutil->error("Malformed event ID", $event) if ($event !~ /^20[0-9]{2}[a-zA-Z0-9\-]+$/);
-$webutil->error("Missing schedule") if (!$schedule);
 $webutil->error("Malformed start", $start) if ($start !~ /^20[0-9]{2}\-[0-9]{2}\-[0-9]{2}$/);
 $webutil->error("Malformed end", $end) if ($end !~ /^20[0-9]{2}\-[0-9]{2}\-[0-9]{2}$/);
 $name = &escapeCsv($name);
 $location = &escapeCsv($location);
 $schedule =~ s/\r\n|\r/\n/g;
-$webutil->error("Malformed CSV", $schedule) if (!$schedule or $schedule !~ /\AMatch,R1,R2,R3,B1,B2,B3\n(?:(?:pm|qm|qf|sf|([1-5]p)|f)[0-9]+(?:,[0-9]+){6}\n)+\Z/g);
+$webutil->error("Malformed CSV", $schedule) if ($schedule !~ /\A(Match,R1,R2,R3,B1,B2,B3\n(?:(?:pm|qm|qf|sf|([1-5]p)|f)[0-9]+(?:,[0-9]+){6}\n)+)?\Z/g);
 
 my $dbh = $db->dbConnection();
 
 sub writeCsvData(){
-	my $file = "../data/${event}.schedule.csv";
-	my $lockFile = "$file.lock";
-	open(my $lock, '>', $lockFile) or $webutil->error("Cannot open $lockFile", "$!\n");
-	flock($lock, LOCK_EX) or $webutil->error("Cannot lock $lockFile", "$!\n");
-	if ( -e $file){
-		my $oldSchedule = read_file($file, {binmode => ':encoding(UTF-8)'});
-		my ($oldPractice) = $oldSchedule =~ /((?:^pm.*\n)+)/m;
-		my ($oldQuals) = $oldSchedule =~ /((?:^qm.*\n)+)/m;
-		my ($oldPlayoffs) = $oldSchedule =~ /((?:^(?:qf|sf|(?:[1-5]p)).*\n)+)/m;
-		my ($headers) = $schedule =~ /((?:^Match.*\n)+)/m;
-		my ($newPractice) = $schedule =~ /((?:^pm.*\n)+)/m;
-		my ($newQuals) = $schedule =~ /((?:^qm.*\n)+)/m;
-		my ($newPlayoffs) = $schedule =~ /((?:^(?:qf|sf|(?:[1-5]p)).*\n)+)/m;
-		$schedule = $headers.($newPractice||$oldPractice||"").($newQuals||$oldQuals||"").($newPlayoffs||$oldPlayoffs||"");
+	my ($file, $lockFile, $lock);
+	if ($schedule){
+		$file = "../data/${event}.schedule.csv";
+		$lockFile = "$file.lock";
+		open($lock, '>', $lockFile) or $webutil->error("Cannot open $lockFile", "$!\n");
+		flock($lock, LOCK_EX) or $webutil->error("Cannot lock $lockFile", "$!\n");
+		if ( -e $file){
+			my $oldSchedule = read_file($file, {binmode => ':encoding(UTF-8)'});
+			my ($oldPractice) = $oldSchedule =~ /((?:^pm.*\n)+)/m;
+			my ($oldQuals) = $oldSchedule =~ /((?:^qm.*\n)+)/m;
+			my ($oldPlayoffs) = $oldSchedule =~ /((?:^(?:qf|sf|(?:[1-5]p)).*\n)+)/m;
+			my ($headers) = $schedule =~ /((?:^Match.*\n)+)/m;
+			my ($newPractice) = $schedule =~ /((?:^pm.*\n)+)/m;
+			my ($newQuals) = $schedule =~ /((?:^qm.*\n)+)/m;
+			my ($newPlayoffs) = $schedule =~ /((?:^(?:qf|sf|(?:[1-5]p)).*\n)+)/m;
+			$schedule = $headers.($newPractice||$oldPractice||"").($newQuals||$oldQuals||"").($newPlayoffs||$oldPlayoffs||"");
+		}
 	}
 
 	$webutil->error("Error opening $file for writing", "$!") if (!open my $fh, ">", $file);
@@ -103,20 +105,22 @@ sub writeDbData(){
 		$blueAllianceId = $data->[0]->[0]||$blueAllianceId;
 		$firstInspiresId = $data->[0]->[1]||$firstInspiresId;
 	}
-	if ($schedule =~ /((?:^pm.*\n)+)/m){
-		$dbh->prepare("DELETE FROM `schedule` WHERE `site`=? AND `event`=? AND `match` like 'pm%'")->execute($db->getSite(), $event);
-	}
-	if ($schedule =~ /((?:^qm.*\n)+)/m){
-		$dbh->prepare("DELETE FROM `schedule` WHERE `site`=? AND `event`=? AND `match` like 'qm%'")->execute($db->getSite(), $event);
-	}
-	if ($schedule =~ /((?:^(?:qf|sf|(?:[1-5]p)).*\n)+)/m){
-		$dbh->prepare("DELETE FROM `schedule` WHERE `site`=? AND `event`=? AND (`match` like 'qf%' OR `match` like 'sf%' OR `match` like '1p%' OR `match` like '2p%' OR `match` like '3p%' OR `match` like '4p%' OR `match` like '5p%')")->execute($db->getSite(), $event);
-	}
-	my $csv = csv->new($schedule);
-	for my $row (1..$csv->getRowCount()){
-		my $data = $csv->getRowMap($row);
-		$data->{'event'}=$event;
-		$db->upsert('schedule', $data);
+	if ($schedule){
+		if ($schedule =~ /((?:^pm.*\n)+)/m){
+			$dbh->prepare("DELETE FROM `schedule` WHERE `site`=? AND `event`=? AND `match` like 'pm%'")->execute($db->getSite(), $event);
+		}
+		if ($schedule =~ /((?:^qm.*\n)+)/m){
+			$dbh->prepare("DELETE FROM `schedule` WHERE `site`=? AND `event`=? AND `match` like 'qm%'")->execute($db->getSite(), $event);
+		}
+		if ($schedule =~ /((?:^(?:qf|sf|(?:[1-5]p)).*\n)+)/m){
+			$dbh->prepare("DELETE FROM `schedule` WHERE `site`=? AND `event`=? AND (`match` like 'qf%' OR `match` like 'sf%' OR `match` like '1p%' OR `match` like '2p%' OR `match` like '3p%' OR `match` like '4p%' OR `match` like '5p%')")->execute($db->getSite(), $event);
+		}
+		my $csv = csv->new($schedule);
+		for my $row (1..$csv->getRowCount()){
+			my $data = $csv->getRowMap($row);
+			$data->{'event'}=$event;
+			$db->upsert('schedule', $data);
+		}
 	}
 	$db->upsert('event', {
 		'event' => $event,
